Move scala native stuff to cli module
